+++ MSAL/test/unit/MSALPublicClientApplicationTests.m
@@ -86,9 +86,13 @@
 #import "MSIDDeviceInfo.h"
 #import "MSALTestCacheTokenResponse.h"
 #import "MSALAuthenticationSchemePop.h"
+#import "MSALWipeCacheForAllAccountsConfig.h"
 
 #if TARGET_OS_IPHONE
 #import "MSIDApplicationTestUtil.h"
+#else
+#import "MSALAccountsProvider.h"
+#import "MSIDMacACLKeychainAccessor.h"
 #endif
 
 #pragma clang diagnostic push
@@ -3470,7 +3474,177 @@
     [self waitForExpectations:@[expectation] timeout:1];
 }
 
+- (void)testSignoutWithAccount_whenNonNilAccount_andWipeAllAccountsTrue_andBrokerDisabled_shouldWipeAllAccounts
+{
+    // Save default response
+    [self msalStoreTokenResponseInCache];
+    
+    // Save tokens for the same account for a different client
+    [self msalStoreSecondAppTokenResponseInCache];
+    
+    MSIDAccount *account = [[MSIDAADV2Oauth2Factory new] accountFromResponse:[self msalDefaultTokenResponse]
+                                                               configuration:[self msalDefaultConfiguration]];
+    MSALAccount *msalAccount = [[MSALAccount alloc] initWithMSIDAccount:account createTenantProfile:NO];
+        
+    NSError *error = nil;
+    __auto_type authority = [@"https://login.microsoftonline.com/common" msalAuthority];
+    
+    MSALPublicClientApplicationConfig *config = [[MSALPublicClientApplicationConfig alloc] initWithClientId:UNIT_TEST_CLIENT_ID
+                                                                                                redirectUri:nil
+                                                                                                  authority:authority];
+    MSALPublicClientApplication *application = [[MSALPublicClientApplication alloc] initWithConfiguration:config
+                                                                                                    error:&error];
+    
+    XCTAssertNotNil(application);
+    XCTAssertNil(error);
+    
+    MSALPublicClientApplication *secondApplication = [self createSecondTestAppWithAuthority:authority];
+    XCTAssertNotNil(secondApplication);
+    
+    MSALWebviewParameters *webParams = [[MSALWebviewParameters alloc] initWithAuthPresentationViewController:[self.class sharedViewControllerStub]];
+    MSALSignoutParameters *parameters = [[MSALSignoutParameters alloc] initWithWebviewParameters:webParams];
+    parameters.wipeCacheForAllAccounts = YES;
+    MSALGlobalConfig.brokerAvailability = MSALBrokeredAvailabilityNone;
+    
+    XCTAssertEqual([application allAccounts:nil].count, 1);
+    XCTAssertEqual([secondApplication allAccounts:nil].count, 1);
+    
+    XCTestExpectation *expectation = [self expectationWithDescription:@"Signout"];
+    
+    [application signoutWithAccount:msalAccount
+                  signoutParameters:parameters completionBlock:^(BOOL success, NSError * _Nullable error) {
+        
+        XCTAssertTrue(success);
+        XCTAssertNil(error);
+
+        // All accounts from all clients were wiped
+        XCTAssertEqual([application allAccounts:nil].count, 0);
+        XCTAssertEqual([secondApplication allAccounts:nil].count, 0);
+        
+        MSALAccountEnumerationParameters *accountEnumerationOptions = [MSALAccountEnumerationParameters new];
+        accountEnumerationOptions.returnOnlySignedInAccounts = NO;
+        
+        // All accounts from all clients were wiped
+        NSArray *firstAppSignedOutAccounts = [application accountsForParameters:accountEnumerationOptions error:nil];
+        XCTAssertEqual([firstAppSignedOutAccounts count], 0);
+        
+        // All accounts from all clients were wiped
+        NSArray *secondAppSignedOutAccounts = [application accountsForParameters:accountEnumerationOptions error:nil];
+        XCTAssertEqual([secondAppSignedOutAccounts count], 0);
+        
+        [expectation fulfill];
+    }];
+    
+    [self waitForExpectations:@[expectation] timeout:1];
+}
+
 #endif
+
+- (void)testSignoutWithAccount_whenNonNilaccount_andWipeAllAccountsTrue_andBrokerDisabled_shouldWipeAllAdditionalPartnerLocations
+{
+    [self msalStoreTokenResponseInCache];
+    
+    MSIDAccount *account = [[MSIDAADV2Oauth2Factory new] accountFromResponse:[self msalDefaultTokenResponse]
+                                                               configuration:[self msalDefaultConfiguration]];
+    MSALAccount *msalAccount = [[MSALAccount alloc] initWithMSIDAccount:account createTenantProfile:NO];
+        
+    NSError *error = nil;
+    __auto_type authority = [@"https://login.microsoftonline.com/common" msalAuthority];
+    
+    MSALPublicClientApplicationConfig *config = [[MSALPublicClientApplicationConfig alloc] initWithClientId:UNIT_TEST_CLIENT_ID
+                                                                                                redirectUri:nil
+                                                                                                  authority:authority];
+    MSALPublicClientApplication *application = [[MSALPublicClientApplication alloc] initWithConfiguration:config
+                                                                                                    error:&error];
+    
+    XCTAssertNotNil(application);
+    XCTAssertNil(error);
+    
+    [MSIDTestSwizzle classMethod:@selector(additionalPartnerLocations)
+                              class:[MSALWipeCacheForAllAccountsConfig class]
+                              block:(id)^(void)
+     {
+        return @{
+            @"Microsoft Office Test" : @{(id)kSecAttrAccount : @"test-account-office",
+                                      (id)kSecAttrLabel : @"test-label-office",
+                                    (id)kSecAttrService : @"test-service-office"},
+
+            @"Microsoft Teams Test" : @{(id)kSecAttrAccount : @"test-account-teams",
+                                     (id)kSecAttrLabel : @"test-label-teams",
+                                   (id)kSecAttrService : @"test-service-teams"}
+        };
+     }];
+    
+#if !TARGET_OS_IPHONE
+    NSMutableArray<MSALAccount *> *accounts = [[NSMutableArray alloc] initWithArray:@[msalAccount]];
+    //swizzle interactive method - MacOS test app doesn't have entitlements that support keychain access group.
+    //adding a host app that has valid entitlements would also require enabling code-signing, which could break CI/CD check
+    //So at the moment, the best approach is to swizzle keychain access APIs
+    [MSIDTestSwizzle instanceMethod:@selector(allAccounts:)
+                              class:[MSALAccountsProvider class]
+                              block:(id)^(void)
+     {
+        return accounts;
+     }];
+    [MSIDTestSwizzle instanceMethod:@selector(removeCredentialsWithQuery:context:error:)
+                              class:[MSIDAccountCredentialCache class]
+                              block:(id)^(void)
+     {
+        [accounts removeAllObjects];
+        return YES;
+     }];
+    [MSIDTestSwizzle instanceMethod:@selector(updateSignInStateForHomeAccountId:clientId:state:context:error:)
+                              class:[MSIDAccountMetadataCacheAccessor class]
+                              block:(id)^(void)
+     {
+        return YES;
+     }];
+    [MSIDTestSwizzle instanceMethod:@selector(clearWithContext:error:)
+                              class:[MSIDKeychainTokenCache class]
+                              block:(id)^(void)
+     {
+        return YES;
+     }];
+    [MSIDTestSwizzle instanceMethod:@selector(getDataWithAttributes:context:error:)
+                              class:[MSIDMacACLKeychainAccessor class]
+                              block:(id)^(void)
+     {
+        // Returned data is not important in this case
+        return nil;
+     }];
+    [MSIDTestSwizzle instanceMethod:@selector(removeItemWithAttributes:context:error:)
+                              class:[MSIDMacACLKeychainAccessor class]
+                              block:(id)^(void)
+     {
+        return YES;
+     }];
+#endif
+    
+#if TARGET_OS_IPHONE
+    MSALWebviewParameters *webParams = [[MSALWebviewParameters alloc] initWithAuthPresentationViewController:[self.class sharedViewControllerStub]];
+#else
+    MSALWebviewParameters *webParams = [MSALWebviewParameters new];
+#endif
+    MSALSignoutParameters *parameters = [[MSALSignoutParameters alloc] initWithWebviewParameters:webParams];
+    parameters.wipeCacheForAllAccounts = YES;
+    MSALGlobalConfig.brokerAvailability = MSALBrokeredAvailabilityNone;
+    
+    XCTAssertEqual([application allAccounts:nil].count, 1);
+    
+    XCTestExpectation *expectation = [self expectationWithDescription:@"Signout"];
+    
+    [application signoutWithAccount:msalAccount
+                  signoutParameters:parameters completionBlock:^(BOOL success, NSError * _Nullable error) {
+        
+        XCTAssertTrue(success);
+        XCTAssertNil(error);
+
+        XCTAssertEqual([application allAccounts:nil].count, 0);
+        [expectation fulfill];
+    }];
+    
+    [self waitForExpectations:@[expectation] timeout:1];
+}
 
 - (void)testInitWithConfiguration_WhenBypassRedirectURIIsDefault_ShouldBlockInvalidURI
 {