+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [1.2.0]
 * Multi-tenant PkeyAuth support in MSAL (#1438)
+* Add support to wipe cache for all accounts (#1426)
 
 ## [1.1.26]
 * Added more string utils in common core (#1417)